+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- GraphQL mutation is now imported from `checkout-resources`.
+
 ## [0.2.0] - 2019-08-29
 
 ### Changed